+++ lib/gpui/native/generated.ex
@@ -2,7 +2,7 @@ defmodule GPUI.Native.Generated do
   @moduledoc false
   defmacro __using__(_opts) do
     quote do
-      def start_runtime() do
+      def start_runtime do
         :erlang.nif_error(:nif_not_loaded)
       end